+++ core/src/lib.rs
@@ -95,7 +95,7 @@ pub use serde_json::{
 pub use std::borrow::Cow;
 
 /// Ten megabytes.
-pub const TEN_MB_SIZE_BYTES: u32 = 10 * 1024 * 1024;
+pub const TEN_MB_SIZE_BYTES: usize = 10 * 1024 * 1024;
 
 /// The return type if the subscription wants to return `Result`.
 pub type SubscriptionResult = Result<(), StringError>;
